LayoutTests/http/tests/webgpu/webgpu/web_platform/canvas/configure-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Query `webgpu:web_platform,canvas,configure:*` does not match any cases

Harness Error (FAIL), message = Unhandled rejection: Query `webgpu:web_platform,canvas,configure:*` does not match any cases
